api/requirements.txt
Pillow==7.2.0
pypng==0.0.20
PyQRCode==1.2.1
zbarlight==2.3
requests==2.24.0